Add context-aware linting with filename support
This pull request adds a context to the linter rule to be able to distinguish between file linting and snippet linting. Rules can now access filename information through a shared context object, enabling different behavior for actual files versus code snippets used in documentation or testing.

We added this functionality so we could remove the `erb-requires-trailing-newline` offesnes in all the code snippets in the docs. But this context also allows us in the future to pass around context to build more sophisticated rules.