Language Server: Add configurable `excludedRules` to prevent duplicate diagnostics
This pull request adds configurable `excludedRules` support to the language server and VSCode extension.

Since #362 introduced the `parser-no-errors` rule we now have two ways of surfcing the parser errors in the language server.

The language server now allows to configure which linter rules to exclude from diagnostics. The `parser-no-errors` rule is set as the default to prevent duplicate parsers errors. But this change now also allows users to exclude other linter rules they might want to disable (just in the language server for now).