Document an additional behavior of erb-righ-trim
erb-right-trim flagged a number of instances where we're using `-%>`
when we should be using `%>`. In fixing them, I'd like to be able to
link my team to the herb docs where they can read more about the change,
but the page doesn't currently mention this behavior of this linter.
Seems worth mentioning.