C: New `hb_array` implementation (#782)
This PR adds a new version of the `hb_array` called `hb_narray` that
implements an array that can actually hold the elements it stores.

The existing `hb_array` only allows to store pointers to heap allocated
elements, the new `hb_narray` acts as a container that also provides the
backing memory for the elements it holds.

It's important to note that you can still make the elements pointers if
you want to.

## Why? 

- Improves locality of the elements
- Reduces the number of individual heap allocations for elements

## Next steps

1. Get this merged
2. Refactor existing `hb_array` usages
3. Remove `hb_array` and rename `hb_narray` to `hb_array` 

## Memory layout


![herb_new_array](https://github.com/user-attachments/assets/8012ff75-38c0-418c-a188-6a8d0fa90baa)