C: Revert lazy array initialization and restore `hb_array_size()` usage (#1025)
This pull request reverts #885 which caused memory issues in WebAssembly
by always allocating arrays, even when `NULL` was passed.

The `hb_array_size()` function safely returns `0` for `NULL` arrays,
eliminating the need for allocating arrays for all AST Nodes. This was
causing the `"Cannot enlarge memory"` errors in WASM builds, as seen in
#1016.

This pull request brings back `hb_array_size()` in places where we
previously accessed `array->size` directly. With that, we don't need to
allocate `array_init(8)` just so we can assume we always have a `size`.

The problem is, even if we would use `array_init(0)` it would allocate
memory that we are not going to need. So for now, I think it's better to
be avoid allocating arrays and use `hb_array_size()` until we have fully
integrated the new array implementation from #782. (/cc @timkaechele).

I also added a `hb_narray_size()` for API consistency that should help
with the future array migration.

Resolves #1016